requirements.txt
aliyun-python-sdk-core==2.13.10
aliyun-python-sdk-imm==1.15.1
amqp==2.5.1
beautifulsoup4==4.8.1
billiard==3.6.1.0
bs4==0.0.1
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
Django==2.1.8
django-cors-headers==3.1.0
django-filter==2.2.0
django-haystack==2.8.1
django-redis==4.10.0
django-rest-elasticsearch==0.4.1
djangorestframework==3.10.3
dnspython==1.16.0
drf-haystack==1.8.6
dwebsocket==0.5.11
elasticsearch==2.4.1
elasticsearch-dsl==5.1.0
eventlet==0.25.1
greenlet==0.4.15
idna==2.8
importlib-metadata==0.23
jmespath==0.9.4
kombu==4.6.5
mock==3.0.5
monotonic==1.5
more-itertools==7.2.0
pbkdf2==1.3
pycryptodome==3.9.0
pyDes==2.0.1
PyJWT==1.7.1
PyMySQL==0.9.3
pytest-runner==5.1
python-dateutil==2.8.0
pytz==2019.2
redis==3.3.8
requests==2.22.0
six==1.12.0
soupsieve==1.9.4
urllib3==1.25.5
vine==1.3.0
zipp==0.6.0
